put dataset into main function
